src/requirements.txt
aws-lambda-typing==2.20.0
bcrypt==4.2.1
boto3==1.36.6
botocore==1.36.6
botocore-stubs==1.36.6
jmespath==1.0.1
mypy-boto3-dynamodb==1.36.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
types-awscrt==0.23.7
types-boto3==1.36.6
types-s3transfer==0.11.2
urllib3==2.3.0
